+++ doc/changelog.md
@@ -9,6 +9,7 @@
 - Added `auxiliary` code.
 - Added `base` library support for UTF-8.
 - Added `base` library support for `atoi`-like string conversion.
+- Added `choose` and `choosy`.
 - Added `doc/logo`.
 - Added `endwhile` syntax.
 - Added `example/cbor-to-json`.